+++ src/services/fetchers/PolygonIOCurrencySnapshotGramsFetcher.ts
@@ -33,7 +33,6 @@ export class PolygonIOCurrencySnapshotGramsFetcher
   private pIOCurrencySnapshotGramsDataRepository!: PolygonIOCurrencySnapshotGramsDataRepository;
   @inject(PriceDataRepository) private priceDataRepository!: PriceDataRepository;
 
-  private logPrefix = `[${FetcherName.PolygonIOCurrencySnapshotGrams}]`;
   static fetcherSource = '';
 
   constructor(@inject('Settings') settings: Settings) {
@@ -41,6 +40,8 @@ export class PolygonIOCurrencySnapshotGramsFetcher
     this.apiKey = settings.api.polygonIO.apiKey;
     this.timeout = settings.api.polygonIO.timeout;
     this.valuePath = '$.ticker.lastQuote.a';
+
+    this.logPrefix = `[${FetcherName.PolygonIOCurrencySnapshotGrams}]`;
   }
 
   async apply(
@@ -50,7 +51,7 @@ export class PolygonIOCurrencySnapshotGramsFetcher
     try {
       await this.fetchPrices(params);
     } catch (e) {
-      this.logger.error(`${this.logPrefix} fetchPrices: ${(e as Error).message}`);
+      this.logger.error(`${this.logPrefix} failed: ${(e as Error).message}`);
     }
 
     const {symbols, timestamp} = options;
@@ -107,7 +108,7 @@ export class PolygonIOCurrencySnapshotGramsFetcher
         }
 
         this.logger.debug(`${this.logPrefix}#${ix} price for ${params.ticker} (${ticker.ticker}): ${price}`);
-        return {price, ticker: ticker.ticker, timestamp: ticker.lastQuote.t / 1e3};
+        return {price, ticker: ticker.ticker, timestamp: Math.trunc(ticker.lastQuote.t / 1e3)};
       })
       .filter((d) => d !== undefined) as ParsedResponse[];
   }