+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- new `PolygonIOStockSnapshotPriceFetcher`
+
+### Removed 
+- removed stock fetcher service
+- remove unused fetcher name: `PolygonIOPrice`
 
 ## [8.1.3] - 2024-08-22
 ### Fixed